Add anomaly score to token contract bot
